Type-stable PanelMatrix creation
Making the final padding an Int instead of StaticInteger, the type
of a PanelMatrix does not depend on it's size, so a normal Matrix
can be converted in a type-stable manner.

However, this requires mul! to be compiled for all possible padding
sizes, so JIT compilation takes a long time.